fix bundle transactions and reset params on selection change
